added a Main.js container component to fix npm tests
